+++ src/includes/helpers/class-alma-settings-helper.php
@@ -64,7 +64,6 @@ class Alma_Settings_Helper {
 			'debug'                                      => 'yes',
 			'keys_validity'                              => 'no',
 			'display_in_page'                            => 'no',
-			'inpage_allowed'                             => 'no',
 		);
 	}