+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* feature : In Page checkout for all merchants
 * change : PHP client version 1.11.2
 
 v5.0.3