+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 Changelog
 =========
 
+v5.0.4
+------
 * feat: compatibility Woocommerce 8.1.1
 * change : PHP client version 1.11.2
+* fix: compatibility with WC_Checkout::process_checkout()
 
 v5.0.3
 ------